refactor(public): Removing unwanted assets
