+++ index.html
@@ -2,7 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + TS</title>
     <div>JSON</div>
@@ -11,9 +10,9 @@
     <div id="jsonlogic"></div>
   </head>
   <body>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="./src/main.ts"></script>
     <script type="module">
-      import { transformToJSON, transformToJSONLogic } from '/src/main.ts';
+      import { transformToJSON, transformToJSONLogic } from './src/main.ts';
 
       const expression = {
         and: [
